+++ runtime/auth/aws-signing-common/build.gradle.kts
@@ -14,6 +14,7 @@ kotlin {
         commonMain {
             dependencies {
                 api(project(":runtime:auth:aws-credentials"))
+                api(project(":runtime:auth:http-auth"))
                 api(project(":runtime:protocol:http"))
 
                 // FIXME - this seems off, (Endpoint is consumed by presigner)
