+++ runtime/auth/http-auth/common/src/aws/smithy/kotlin/runtime/auth/HttpSigner.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package aws.smithy.kotlin.runtime.http.auth
+package aws.smithy.kotlin.runtime.auth
 
 import aws.smithy.kotlin.runtime.InternalApi
 import aws.smithy.kotlin.runtime.client.ExecutionContext
