+++ settings.gradle.kts
@@ -38,6 +38,7 @@ include(":runtime:auth:aws-signing-common")
 include(":runtime:auth:aws-signing-crt")
 include(":runtime:auth:aws-signing-default")
 include(":runtime:auth:aws-signing-tests")
+include(":runtime:auth:http-auth")
 include(":runtime:crt-util")
 include(":runtime:logging")
 include(":runtime:protocol:http")
